R6 class representing a question

This class represents a question. It has a title, a text, a category and a type.

Active bindings

list

returns the question as a list

Methods

Public methods

Question$new()

Question$import_gift()

Question$export_gift()

Question$import()

Question$print()

Question$clone()

Method new()

This function initializes the question with a title

Usage

Question$new(data)

Arguments

data

A list containing the data of the question

Returns

The question object

Method import_gift()

This function imports a question from a string

Usage

Question$import_gift(string)

Arguments

string

A string containing the question data in GIFT format

Returns

The question object

Method export_gift()

This function returns the question as a list

Usage

Question$export_gift()

Returns

A list representation of the question

Method import()

This function imports a question from a list

Usage

Question$import(data)

Arguments

data

A list containing the question data

Returns

The question object

Method print()

This function prints the question

Usage

Question$print()

Method clone()

The objects of this class are cloneable with this method.

Usage

Question$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.